fs24-mc2-etiiiR/requirements.txt
accelerate==0.30.1
accessible-pygments==0.0.4
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
azure-core==1.30.1
azure-identity==1.16.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confluent-kafka==2.3.0
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.5
cupy-cuda12x==13.1.0
cycler==0.12.1
dataclasses-json==0.6.5
datasets==2.19.1
datasketch==1.6.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
docker==7.1.0
docutils==0.20.1
dulwich==0.21.7
eclipse-zenoh==0.10.1rc0
entrypoints==0.4
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.1
fastjsonschema==2.19.1
fastrlock==0.8.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gensim==4.3.2
google-auth==2.29.0
googleapis-common-protos==1.63.0
gpt4all==2.4.0
graphviz==0.20.2
greenlet==3.0.3
grpcio==1.63.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imagesize==1.4.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
installer==0.7.0
ipykernel==6.29.3
ipython==8.24.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
itables==2.1.0
jaraco.classes==3.3.1
jax==0.4.28
jaxlib==0.4.28
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-book==1.0.0
jupyter-cache==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_nvdashboard==0.11.0
jupyterlab_nvidia_nsight==0.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
jupytext==1.16.1
kaggle==1.6.12
keyring==24.3.1
kiwisolver==1.4.5
kubernetes==29.0.0
latexcodec==3.0.0
lazy_loader==0.4
lightfm==1.17
lime==0.2.0.1
linkify-it-py==2.0.3
llvmlite==0.42.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
minijinja==2.0.1
mistune==0.8.4
ml-dtypes==0.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
myst-nb==1.0.0
myst-parser==2.0.0
nbclient==0.10.0
nbconvert==6.5.4
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.24.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.20.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.0
portalocker==2.8.2
posthog==3.5.0
pretty-jupyter==2.0.7
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydata-sphinx-theme==0.15.2
Pygments==2.17.2
PyJWT==2.8.0
pynvim==0.5.0
pynvml==11.5.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-highcharts==0.4.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.6.2
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.23.1
scikit-learn==1.4.2
scikit-surprise==1.1.3
scipy==1.12.0
seaborn==0.13.2
selenium==4.19.0
Send2Trash==1.8.2
sentence-transformers==2.7.0
shap==0.45.1
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.4
snakeviz==2.2.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-book-theme==1.1.2
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==1.0.0
sphinx-multitoc-numbering==0.1.3
sphinx-thebe==0.3.1
sphinx-togglebutton==0.3.2
sphinx_design==0.5.0
sphinx_external_toc==1.0.1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.1
striprtf==0.0.26
surrealdb==0.3.2
sweetviz==2.3.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.4.0
tifffile==2024.4.18
tiktoken==0.6.0
tinycss2==1.2.1
tmdbv3api==1.9.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.4
torch==2.2.2+cu121
torch_geometric==2.5.3
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2024.3.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
widgetsnbextension==4.0.10
windows-curses==2.3.3
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
